+++ clx_retail_pricelist/RELEASE_NOTES.md
@@ -4,3 +4,8 @@
 ### Version 13.0.0.0.2
 #### Setup Module
 - [IMP] CLX Retail Pricelist: Code cleanup
+
+## 07/23/2020
+### Version 13.1.0.0.3
+#### Customer Pricelist
+- [IMP] CLX Retail Pricelist: set pricelist if customer has 'Billing Contact' linked contact and have parent in linked contact else Public pricelist